# api-gateway/docs/swagger.yaml
basePath: /
definitions:
  models.Athlete:
    properties:
      country_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: integer
      id:
        type: string
      name:
        type: string
      sport_type:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateAthleteRequest:
    properties:
      country_id:
        type: string
      name:
        type: string
      sport_type:
        type: string
    type: object
...
  models.CreateMedalRequest:
    properties:
      athlete_id:
        type: string
      country_id:
        type: string
      event_id:
        type: string
      type:
        $ref: '#/definitions/models.MedalType'
    type: object
...
  models.DeleteAthleteResponse:
    properties:
      status:
        type: string
    type: object
...
  models.DeleteMedalResponse:
    properties:
      success:
        type: boolean
    type: object
...
  models.GetAthleteResponse:
    properties:
      country_id:
        type: string
      country_name:
        type: string
      created_at:
        type: string
      deleted_at:
        type: integer
      id:
        type: string
      name:
        type: string
      sport_type:
        type: string
      updated_at:
        type: string
    type: object
  models.GetMedalByFilterRequest:
    properties:
      athlete_id:
        type: string
      country_id:
        type: string
      event_id:
        type: string
      type:
        $ref: '#/definitions/models.MedalType'
    type: object
  models.GetMedalByFilterResponse:
    properties:
      medals:
        items:
          $ref: '#/definitions/models.Medal'
        type: array
    type: object
  models.GetMedalByIdResponse:
    properties:
      athlete_id:
        type: string
      country_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: integer
      event_id:
        type: string
      id:
        type: string
      type:
        $ref: '#/definitions/models.MedalType'
      updated_at:
        type: string
    type: object
  models.GetMedalsResponse:
    properties:
      medals:
        items:
          $ref: '#/definitions/models.Medal'
        type: array
    type: object
...
  models.ListOfAthleteResponse:
    properties:
      athletes:
        items:
          $ref: '#/definitions/models.GetAthleteResponse'
        type: array
    type: object
...
  models.Medal:
    properties:
      athlete_id:
        type: string
      country_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: integer
      event_id:
        type: string
      id:
        type: string
      type:
        $ref: '#/definitions/models.MedalType'
      updated_at:
        type: string
    type: object
  models.MedalType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - GOLD
    - SILVER
    - BRONZE
  models.Message:
    properties:
      err:
        type: string
    type: object
...
  models.UpdateAthleteRequest:
    properties:
      country_id:
        type: string
      id:
        type: string
      name:
        type: string
      sport_type:
        type: string
    type: object
...
  models.UpdateMedalRequest:
    properties:
      athlete_id:
        type: string
      country_id:
        type: string
      event_id:
        type: string
      id:
        type: string
      type:
        $ref: '#/definitions/models.MedalType'
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: Api-gateway service
  version: "1.0"
paths:
  /athletes:
    get:
      consumes:
      - application/json
      description: This method gets athletes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListOfAthleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: GET ATHLETES
      tags:
      - ATHLETE
    post:
      consumes:
      - application/json
      description: This method creates athlete
      parameters:
      - description: Athlete
        in: body
        name: athlete
        required: true
        schema:
          $ref: '#/definitions/models.CreateAthleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Athlete'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: CREATE ATHLETE
      tags:
      - ATHLETE
  /athletes/{id}:
    delete:
      consumes:
      - application/json
      description: This method deletes an athlete
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteAthleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: DELETE ATHLETE
      tags:
      - ATHLETE
    get:
      consumes:
      - application/json
      description: This method gets athlete
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAthleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: GET ATHLETE
      tags:
      - ATHLETE
    put:
      consumes:
      - application/json
      description: This method updates athlete
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Athlete
        in: body
        name: athlete
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAthleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Athlete'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: UPDATE ATHLETE
      tags:
      - ATHLETE
...
  /medal:
    post:
      consumes:
      - application/json
      description: This method creates a medal
      parameters:
      - description: Medal
        in: body
        name: medal
        required: true
        schema:
          $ref: '#/definitions/models.CreateMedalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Medal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: CREATE MEDAL
      tags:
      - MEDAL
  /medal/{id}:
    delete:
      consumes:
      - application/json
      description: This method deletes a medal
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteMedalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: DELETE MEDAL
      tags:
      - MEDAL
    get:
      consumes:
      - application/json
      description: This method gets a medal by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetMedalByIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: GET MEDAL BY ID
      tags:
      - MEDAL
    put:
      consumes:
      - application/json
      description: This method updates a medal
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Medal
        in: body
        name: medal
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMedalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Medal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: UPDATE MEDAL
      tags:
      - MEDAL
  /medal/filter:
    post:
      consumes:
      - application/json
      description: This method gets medals by filter
      parameters:
      - description: Filter
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/models.GetMedalByFilterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetMedalByFilterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: GET MEDALS BY FILTER
      tags:
      - MEDAL
  /medals:
    get:
      consumes:
      - application/json
      description: This method gets all medals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetMedalsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - BearerAuth: []
      summary: GET MEDALS
      tags:
      - MEDAL
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authourization
    type: apiKey
swagger: "2.0"